+++ tests/src/Opulence/Framework/Testing/PhpUnit/Http/Mocks/ApplicationTestCase.php
@@ -8,7 +8,6 @@
  */
 namespace Opulence\Tests\Framework\Testing\PhpUnit\Http\Mocks;
 
-use Monolog\Logger;
 use Opulence\Applications\Application;
 use Opulence\Applications\Environments\Environment;
 use Opulence\Bootstrappers\Paths;
@@ -22,7 +21,7 @@ use Opulence\Framework\Bootstrappers\Http\Views\ViewFunctionsBootstrapper;
 use Opulence\Framework\Testing\PhpUnit\Http\ApplicationTestCase as BaseApplicationTestCase;
 use Opulence\Ioc\Container;
 use Opulence\Ioc\IContainer;
-use Opulence\Tests\Applications\Mocks\MonologHandler;
+use Psr\Log\LoggerInterface;
 
 /**
  * Mocks the HTTP application for use in testing
@@ -49,10 +48,8 @@ class ApplicationTestCase extends BaseApplicationTestCase
      */
     protected function getKernelLogger()
     {
-        $logger = new Logger("application");
-        $logger->pushHandler(new MonologHandler());
-
-        return $logger;
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        return $this->getMock(LoggerInterface::class);
     }
 
     /**